split socket concerns
